+++ src/CryptoNoteCore/TransactionValidator.h
@@ -48,7 +48,7 @@ class ValidateTransaction
             const CryptoNote::Currency &currency,
             const CryptoNote::Checkpoints &checkpoints,
             Utilities::ThreadPool<bool> &threadPool,
-            const uint64_t blockHeight,
+            const uint32_t blockHeight,
             const uint64_t blockSizeMedian,
             const uint64_t minFee,
             const bool isPoolTransaction);
@@ -95,7 +95,7 @@ class ValidateTransaction
 
         const CryptoNote::Checkpoints &m_checkpoints;
 
-        const uint64_t m_blockHeight;
+        const uint32_t m_blockHeight;
 
         const uint64_t m_blockSizeMedian;
